Use the built-in Fraction class
Instead of defining our own Ratio class, just use the built-in Fraction
class instead (which means we inherit the ability to use the fraction
with normal arithmetic operators). This modification re-defines the
Ratio class as a subclass of Fraction which adds den and num as
calculated properties for backwards compatibility (the representation is
also overridden to maintain compatibility).