Canon MakerNote: Allow callable to process tag value
While working on some images I found that the library did not expose the 'CameraTemperature' tag from the ShotInfo MakerNote.

To support this I added a callable and allow that callable to be called with the tag value. I might be useful for other tags too.

Example output from a JPEG image taken on a Canon EOS 2000D:
```
>>> exif = exifread.process_file(fd, details=True)
>>> exif["MakerNote CameraTemperature"]
(0x0000) Proprietary=55 C @ 0
```

Signed-off-by: Daan van Gorkum <djvg@djvg.net>